requirements.txt
backports.shutil-get-terminal-size==1.0.0
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
DateTime==4.2
decorator==4.1.2
Django==1.10
djangorestframework==3.7.7
enum34==1.1.6
http-parser==0.8.3
idna==2.6
ipdb==0.10.3
ipython==5.5.0
ipython-genutils==0.2.0
oauthlib==2.0.6
pathlib2==2.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
Pygments==2.2.0
PyJWT==1.5.3
python-decouple==3.1
python-openid==2.2.5
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
restkit==4.2.2
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
socketpool==0.5.3
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
win-unicode-console==0.5
zope.interface==4.4.3
